fixed issue #3211
On page: https://sendgrid.com/docs/User_Guide/Suppressions/invalid_emails.html

Added the following to the end of the first paragraph:

Before you send, SendGrid does a check against the format of the email address to attempt to verify its validity. If the recipient server checks the address and doesn't find it, they will send back a 550 bounce to say that this is an invalid email address.